pages: migrate channel info page
